Switch all world coordinates to unsigned integers. This makes negative coordinates physically impossible, and simplifies further development. It is easier to limit the coordinate space, rather than trying to handle negative coordinates properly.

+++ config/config.go
@@ -11,8 +11,8 @@ const (
 	WorldHeight int64 = 1024
 
 	// Initial player position, when the world is created
-	PlayerStartX float64 = float64(WorldWidth) / 2
-	PlayerStartY float64 = float64(WorldHeight) / 2
+	PlayerStartX int64 = WorldWidth / 2
+	PlayerStartY int64 = WorldHeight / 2
 
 	WorldSaveDirectory string = "./saves/"
 	WorldInfoFile      string = "world.gob"
